:hammer: Refactor : 우승자 함수 분리
